% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toggl-helpers.R
\name{get.toggl.clients}
\alias{get.toggl.clients}
\title{Fetch a list of clients from toggl.}
\usage{
get.toggl.clients(toggl_token, workspace_id, verbose = FALSE)
}
\arguments{
\item{toggl_token}{A toggl API token to access https://toggl.com.}

\item{workspace_id}{The workspace id for the wanted workspace.}

\item{verbose}{A flag to enable more verbose output, Default value: FALSE}
}
\value{
The list (JSON) of clients accessable using \code{toggl_token} from the toggl workspace with id: \code{workspace_id}.
}
\description{
Fetch a list of clients from toggl.
}
\examples{
get.toggl.clients(Sys.getenv("TOGGL_TOKEN"), Sys.getenv("TOGGL_WORKSPACE"))
}
\seealso{
Other get.toggl: \code{\link{get.toggl.groups}}
}
\concept{get.toggl}
