% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toggl-helpers.R
\name{get.toggl.groups}
\alias{get.toggl.groups}
\title{Fetch a list of groups from toggl.}
\usage{
get.toggl.groups(toggl_token, workspace_id, verbose = FALSE)
}
\arguments{
\item{toggl_token}{A toggl API token to access https://toggl.com.}

\item{workspace_id}{The workspace id for the wanted workspace.}

\item{verbose}{A flag to enable more verbose output, Default value: FALSE}
}
\value{
A data frame of the groups accessable using \code{toggl_token} from the toggl workspace with id: \code{workspace_id}.
}
\description{
Fetch a list of groups from toggl.
}
\examples{
get.toggl.groups(Sys.getenv("TOGGL_TOKEN"), Sys.getenv("TOGGL_WORKSPACE"))
}
\seealso{
Other get.toggl: \code{\link{get.toggl.clients}},
  \code{\link{get.toggl.group.data}},
  \code{\link{get.toggl.v2.group.details}},
  \code{\link{get.toggl.v8.data}},
  \code{\link{get.toggl.v8}}
}
\concept{get.toggl}
