% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toggl-helpers.R
\name{get.toggl.v2.group.details}
\alias{get.toggl.v2.group.details}
\title{Fetch a page of entries for a group using the V2 API}
\usage{
get.toggl.v2.group.details(toggl_token, workspace_id, group,
  since = Sys.Date() - 7, until = Sys.Date(), page = 1,
  verbose = FALSE)
}
\arguments{
\item{toggl_token}{A toggl API token to access https://toggl.com.}

\item{workspace_id}{The workspace id for the wanted workspace.}

\item{group}{The group Id for the wanted group}

\item{since}{The start date of the wanted time interval, Default value: \code{Sys.Date() -7}}

\item{until}{The end date of the wanted interval, Default value: \code{Sys.Date()}}

\item{page}{Which page of the data to fetch, Default value: 1}

\item{verbose}{A flag to enable more verbose output, Default value: FALSE}
}
\value{
A httr response object with the paged data for the wanted \code{group}.
}
\description{
Fetch a page of entries for a group using the V2 API
}
\examples{
get.toggl.v2.group.details(Sys.getenv("TOGGL_TOKEN"), Sys.getenv("TOGGL_WORKSPACE"), <groupId>, since = Sys.Date() - 30)
}
\seealso{
Other get.toggl: \code{\link{get.toggl.clients}},
  \code{\link{get.toggl.group.data}},
  \code{\link{get.toggl.groups}},
  \code{\link{get.toggl.v8.data}},
  \code{\link{get.toggl.v8}}
}
\concept{get.toggl}
