% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toggl-helpers.R
\name{get.toggl.v8}
\alias{get.toggl.v8}
\title{Fetch a list from toggl using the v8 api.}
\usage{
get.toggl.v8(toggl_token, workspace_id, what, verbose = FALSE)
}
\arguments{
\item{toggl_token}{A toggl API token to access https://toggl.com.}

\item{workspace_id}{The workspace id for the wanted workspace.}

\item{what}{What to fetch}

\item{verbose}{A flag to enable more verbose output, Default value: FALSE}
}
\value{
The list (JSON) of clients accessable using \code{toggl_token} from the toggl workspace with id: \code{workspace_id}.
}
\description{
Fetch a list from toggl using the v8 api.
}
\examples{
get.toggl.v8(Sys.getenv("TOGGL_TOKEN"), Sys.getenv("TOGGL_WORKSPACE"), "clients")
}
\seealso{
Other get.toggl: \code{\link{get.toggl.clients}},
  \code{\link{get.toggl.groups}},
  \code{\link{get.toggl.v8.data}}
}
\concept{get.toggl}
