% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toggl-helpers.R
\name{get.toggl.group.data}
\alias{get.toggl.group.data}
\title{Fetch the detailed entries for a group during a specified time interval}
\usage{
get.toggl.group.data(toggl_token, workspace_id, group, since = Sys.Date()
  - 7, verbose = FALSE)
}
\arguments{
\item{toggl_token}{A toggl API token to access https://toggl.com.}

\item{workspace_id}{The workspace id for the wanted workspace.}

\item{group}{The group Id for the wanted group}

\item{since}{The start date of the wanted time interval, Default value: \code{Sys.Date() -7}}

\item{verbose}{A flag to enable more verbose output, Default value: FALSE}

\item{until}{The end date of the wanted interval, Default value: \code{Sys.Date()}}
}
\value{
A data frame with the collected entries for the wanted \code{group}.
}
\description{
Fetch the detailed entries for a group during a specified time interval
}
\examples{
get.toggl.group.data(Sys.getenv("TOGGL_TOKEN"), Sys.getenv("TOGGL_WORKSPACE"), <groupId>, since = Sys.Date() - 30)
}
\seealso{
Other get.toggl: \code{\link{get.toggl.clients}},
  \code{\link{get.toggl.groups}},
  \code{\link{get.toggl.v2.group.details}},
  \code{\link{get.toggl.v8.data}},
  \code{\link{get.toggl.v8}}
}
\concept{get.toggl}
